refactor: 유저 타입 분리
