+++ client/src/types/users.d.ts
@@ -1,0 +1,13 @@
+import { ProjectType } from '@/types/project';
+
+export interface UserProfile {
+  index: number;
+  name: string;
+  imageURL: string;
+  job: string;
+  admin: boolean;
+}
+
+export interface UserInfoWithProject extends UserProfile {
+  projects: ProjectType[];
+}
